Fix Joypad Interrupts
Joypad interrupt was always called when a key was maintained, instead of
being called only when a key changes from released to pressed.
This bug caused Pokémon Gold to be unable to pass the title screen, and
animation on the beginning of the game was blocked when pressing a key,
because the emulator called continuously the joypad interrupt at 0x0060.