Improve PPU modes implementation
- Fix Super Mario Land bug where HUD pixels scrolled with the rest of the
background. This was due to an incorrect sequence of PPU modes and when
the actual emulator's corresponding actions happened. The function
draw_scanline used to call scan_OAM, and draw the scanline, all at once,
at the beginning of MODE 2 (the moment when STAT PPU modes flags turns
to 2). It caused a misbehavior whith SML. Now, scan_OAM is called
outside draw_scanline when STAT turns to MODE 2, and draw_scanline at
when STAT turns to MODE 0, because at this point a scanline is
considered drawn and the game can change graphics registers safely.