Reorganize rendering and vblank functions
- Reorganize render_frame_and_vblank_if_needed in two functions (vblank
and render_frame) for improved logic and readability.
- Fix sprites in first scanline for good: issue was caused by the fact
that scan_OAM was called when LY==154 (should not exist, this state is
only used to know when to render a frame and should only be temporary),
and also because is_OAM_scanned was not reset to false at the begining
of a new scanline.